Added the ability to put custom data into the metadata returned to the callback.
This makes writing a bulk processor much easier: if you need to know the file path or URL of the source media, simply put it in the customData option and it will be waiting for you when the callback fires.  If you need to know the record ID of the relevant enrty in your database, put it in the customData option. Simple tools with powerful uses.